update ArchsimLib
This updates to the latest version of ArchsimLib, which includes construction types. It also eliminates the special ArchsimLibData assembly, because that causes conflicts in downstream applications that want to use ArchsimLib4Umi. This isn't a long-term solution - now we have a direct dependency on ArchsimLib4Umi, which isn't the way way want to go - but we'll have to work something out later.